PusatDjango/requirements.txt
Pillow
python-docx
pvlib
pytz
pandas
djangorestframework
